+++ crates/test-ui/ui/msg_send_id_alloc_init_different.rs
@@ -1,11 +1,11 @@
 //! Ensure that `init` returns the same type as given from `alloc`.
 use objc2::rc::{Allocated, Id};
-use objc2::runtime::{NSObject, AnyObject};
+use objc2::runtime::{AnyObject, NSObject};
 use objc2::{class, msg_send_id};
 
 fn main() {
     let cls = class!(NSObject);
-    let obj: Option<Allocated<NSObject>> = unsafe { msg_send_id![cls, alloc] };
+    let obj: Allocated<NSObject> = unsafe { msg_send_id![cls, alloc] };
 
     let _: Id<AnyObject> = unsafe { msg_send_id![obj, init] };
 }