+++ crates/objc2/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `CounterpartOrSelf`.
 * Added new `encode` traits `EncodeReturn`, `EncodeArgument` and
   `EncodeArguments`.
+* Added methods `as_ptr` and `as_mut_ptr` to `Allocated`.
 
 ### Changed
 * **BREAKING**: `AnyClass::verify_sel` now take more well-defined types
@@ -47,6 +48,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   a method.
 * **BREAKING**: Renamed the associated types `Ret` and `Args` on
   `MethodImplementation` to `Return` and `Arguments`.
+* **BREAKING**: Make `rc::Allocated` allowed to be `NULL` internally, such
+  that uses of `Option<Allocated<T>>` is now simply `Allocated<T>`.
 
 ### Deprecated
 * Soft deprecated using `msg_send!` without a comma between arguments (i.e.